Fixed layout (and refactored MobileBookingHeader)
